Extract ModelUpdater from controller.py (#394 Phase 3F)
New ModelUpdater class owns the per-tick model update logic:
- update() — was __update_model (the main update loop)
- _update_pair_model_state() — per-pair scan/status processing
- sync_persist_to_all_builders() — push persist state to model builders

Controller.process() is now:
    self.__pipeline.propagate_exceptions()
    self.__pipeline.cleanup()
    self.__pipeline.step()
    self.__updater.update()

Pipeline's sync_persist_callback wired to updater after construction.
Test patches updated to target ModelUpdater instead of Controller.

controller.py: 764 → 389 lines (-375)
New file: controller/model_updater.py (443 lines)
Cumulative from 1433: -1044 lines (27% of original)
297 tests pass.
